requirements.txt
requests
pandas
python-dotenv
psycopg2-binary
SQLAlchemy
